// src/app/api/api/models/bitacora-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BitacoraResponse {
  accion?: string;
  activo?: boolean;
  bitacoraId?: number;
  detalle?: string;
  entidad?: string;
  fecha?: string;
  fechaCreacion?: string;
  fechaModificacion?: string;
  modulo?: string;
  registroId?: number;
  resultado?: string;
  usuarioId?: number;
}